requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
asyncpg==0.29.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28
